+++ security/core/src/main/java/org/seedstack/seed/security/internal/SeedSessionStorageEvaluator.java
@@ -19,6 +19,6 @@ class SeedSessionStorageEvaluator implements SessionStorageEvaluator {
 
     @Override
     public boolean isSessionStorageEnabled(Subject subject) {
-        return securityConfig.session().isEnabled();
+        return securityConfig.sessions().isEnabled();
     }
 }