+++ security/core/src/test/java/org/seedstack/seed/security/internal/SecurityInternalModuleUnitTest.java
@@ -9,6 +9,7 @@ package org.seedstack.seed.security.internal;
 
 import com.google.inject.Injector;
 import com.google.inject.PrivateBinder;
+import com.google.inject.Provider;
 import com.google.inject.TypeLiteral;
 import com.google.inject.binder.AnnotatedBindingBuilder;
 import com.google.inject.binder.AnnotatedElementBuilder;
@@ -17,6 +18,7 @@ import org.apache.shiro.realm.Realm;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.internal.util.reflection.Whitebox;
+import org.seedstack.seed.security.SecurityConfig;
 import org.seedstack.seed.security.internal.realms.ConfigurationRealm;
 
 import java.lang.annotation.Annotation;
@@ -37,7 +39,7 @@ public class SecurityInternalModuleUnitTest {
 
     private SecurityConfigurer securityConfigurer;
 
-    @SuppressWarnings({ "unchecked", "rawtypes" })
+    @SuppressWarnings({"unchecked", "rawtypes"})
     @Before
     public void before() {
         binder = mock(PrivateBinder.class);
@@ -47,7 +49,7 @@ public class SecurityInternalModuleUnitTest {
         AnnotatedElementBuilder aeb = mock(AnnotatedElementBuilder.class);
         when(binder.expose(any(Class.class))).thenReturn(aeb);
         ScopedBindingBuilder sb = mock(ScopedBindingBuilder.class);
-        when(ab.toProvider(any(Class.class))).thenReturn(sb);
+        when(ab.toProvider(any(Provider.class))).thenReturn(sb);
         when(binder.bind(any(TypeLiteral.class))).thenReturn(ab);
         when(binder.skipSources(any(Class.class), any(Class.class))).thenReturn(binder);
         securityConfigurer = mock(SecurityConfigurer.class);
@@ -60,13 +62,14 @@ public class SecurityInternalModuleUnitTest {
         Set<RealmConfiguration> realmConfs = new HashSet<>();
         realmConfs.add(new RealmConfiguration("ConfigurationRealm", ConfigurationRealm.class));
         when(securityConfigurer.getConfigurationRealms()).thenReturn(realmConfs);
+        when(securityConfigurer.getSecurityConfiguration()).thenReturn(new SecurityConfig());
 
         underTest.configure();
     }
 
     @Test
     public void testProvider() {
-        SecurityInternalModule.RealmProvider rp = new SecurityInternalModule.RealmProvider();
+        SecurityInternalModule.RealmProvider rp = new SecurityInternalModule.RealmProvider(new SecurityConfig());
         Injector i = mock(Injector.class);
         ShiroRealmAdapter adapter = new ShiroRealmAdapter();
         when(i.getInstance(ShiroRealmAdapter.class)).thenReturn(adapter);